Move endpoint to separate controller
